@@ -15,7 +15,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.7.0</version>
+        <version>3.0.0</version>
         <relativePath /> <!-- lookup parent from repository -->
     </parent>
     <groupId>com.api</groupId>
@@ -24,8 +24,8 @@
     <name>swotlyzer</name>
     <description>swotlyzer</description>
     <properties>
-        <java.version>11</java.version>
-        <spring-boot.version>2.7.0</spring-boot.version>
+        <java.version>17</java.version>
+        <spring-boot.version>3.0.0</spring-boot.version>
         <spring-fox-swagger2.version>2.9.2</spring-fox-swagger2.version>
         <spring-fox-swagger-ui.version>2.9.2</spring-fox-swagger-ui.version>
         <jsonwebtoken.version>0.9.1</jsonwebtoken.version>
@@ -50,11 +50,11 @@
                 <artifactId>spring-dotenv</artifactId>
                 <version>${spring-dotenv.version}</version>
             </dependency>
-            <!-- https://mvnrepository.com/artifact/javax.servlet/javax.servlet-api -->
+            <!-- https://mvnrepository.com/artifact/jakarta.servlet/jakarta.servlet-api -->
             <dependency>
-                <groupId>javax.servlet</groupId>
-                <artifactId>javax.servlet-api</artifactId>
-                <version>4.0.1</version>
+                <groupId>jakarta.servlet</groupId>
+                <artifactId>jakarta.servlet-api</artifactId>
+                <version>6.0.0</version>
                 <scope>provided</scope>
             </dependency>
 
@@ -197,7 +197,7 @@
             </dependency>
 
             <!--            <dependency>-->
-            <!--                <groupId>javax.xml.bind</groupId>-->
+            <!--                <groupId>jakarta.xml.bind</groupId>-->
             <!--                <artifactId>jaxb-api</artifactId>-->
             <!--                <version>${jaxb-api.version}</version>-->
             <!--            </dependency>-->
