@@ -1,6 +1,6 @@
 package com.microservices.swotlyzer.swot.analysis.service.dtos;
 
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotBlank;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
